web-search-agent/pyproject.toml
[project]
name = "firecrawl-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-google-genai>=2.1.9",
    "langchain-mcp-adapters>=0.1.9",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
]
